Moved robots cfg files to assets/robots
